Dodanie obsługi doctora

+++ app/Http/Controllers/Mood/MoodController.php
@@ -23,116 +23,130 @@ use App\Http\Services\Common;
 use Auth;
 class MoodController extends Controller  {
     public function add(Request $request) {
-        $Mood = new Mood;
-        $Mood->checkAddMoodDate($request);
-        $Mood->checkAddMood($request);
-        if (count($Mood->errors) != 0) {
-            return View("ajax.error")->with("error",$Mood->errors);
-        }
-        else {
-            $id = $Mood->saveMood($request);
-            
-            if (!empty($request->get("idAction"))) {
-                $Mood->saveAction($request,$id);
-            }
-        }
-       //print ("<script>document.getElementById('form2').reset() </script>");
-        
-       
-    }
-    
-    public function edit(Request $request) {
-        $Mood = new Mood;
-        $mood = $Mood->selectMood($request);
-        return View("ajax.editMood")->with("list",$mood)->with("i",$request->get("i"));
-        
-    }
-    
-    public function EditAction(Request $request) {
-        $Mood = new Mood;
-
-            $levelMood = $Mood->checkLevel($request->get("levelMood"),"nastroju");
-            $levelAnxiety = $Mood->checkLevel($request->get("levelAnxiety"),"lęku");
-            $levelNervousness = $Mood->checkLevel($request->get("levelNervousness"),"zdenerowania");
-            $levelStimulation = $Mood->checkLevel($request->get("levelStimulation"),"pobudzenia");
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+            $Mood->checkAddMoodDate($request);
+            $Mood->checkAddMood($request);
             if (count($Mood->errors) != 0) {
                 return View("ajax.error")->with("error",$Mood->errors);
             }
             else {
-                $Mood->updateMood($request);
-                return View("ajax.succes")->with("succes","Pomyslnie zmodyfikowany nastrój");
+                $id = $Mood->saveMood($request);
+
+                if (!empty($request->get("idAction"))) {
+                    $Mood->saveAction($request,$id);
+                }
             }
-            
+           //print ("<script>document.getElementById('form2').reset() </script>");
         
+        }
+    }
+    
+    public function edit(Request $request) {
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+            $mood = $Mood->selectMood($request);
+            return View("ajax.editMood")->with("list",$mood)->with("i",$request->get("i"));
+        }
+    }
+    
+    public function EditAction(Request $request) {
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+
+                $levelMood = $Mood->checkLevel($request->get("levelMood"),"nastroju");
+                $levelAnxiety = $Mood->checkLevel($request->get("levelAnxiety"),"lęku");
+                $levelNervousness = $Mood->checkLevel($request->get("levelNervousness"),"zdenerowania");
+                $levelStimulation = $Mood->checkLevel($request->get("levelStimulation"),"pobudzenia");
+                if (count($Mood->errors) != 0) {
+                    return View("ajax.error")->with("error",$Mood->errors);
+                }
+                else {
+                    $Mood->updateMood($request);
+                    return View("ajax.succes")->with("succes","Pomyslnie zmodyfikowany nastrój");
+                }
+
+        }
     }
 
 
     public function Sleepadd(Request $request) {
-        $Mood = new Mood;
-        $Mood->checkAddSleepDate($request);
-        //$Mood->checkAddMood($request);
-        if (count($Mood->errors) != 0) {
-            return View("ajax.error")->with("error",$Mood->errors);
-        }
-        else {
-            $Mood->saveSleep($request);
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+            $Mood->checkAddSleepDate($request);
+            //$Mood->checkAddMood($request);
+            if (count($Mood->errors) != 0) {
+                return View("ajax.error")->with("error",$Mood->errors);
+            }
+            else {
+                $Mood->saveSleep($request);
+            }
         }
     }
     public function AddDescription(Request $request) {
-        $description = AppMood::showDescription($request->get("id"));
-        return View("ajax.editDescription")->with("description",$description->what_work)->with("idMood",$request->get("id"));
+        if (Auth::User()->type == "user") {
+            $description = AppMood::showDescription($request->get("id"));
+            return View("ajax.editDescription")->with("description",$description->what_work)->with("idMood",$request->get("id"));
+        }
     }
     
     
     public function EditDescription(Request $request) {
-        $Mood = new Mood;
-        $Mood->updateDescription($request);
-        return View("ajax.succes")->with("succes","pomyslnie zmodyfikowano");
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+            $Mood->updateDescription($request);
+            return View("ajax.succes")->with("succes","pomyslnie zmodyfikowano");
+        }
     }
     
     
     public function Actionadd(Request $request) {
-        $Action = new Action;
-        $Action->checkAddActionDate($request);
-        $Action->checkAddAction($request);
-        //$Action->checkAddMood($request);
-        if (count($Action->errors) != 0) {
-            return View("ajax.error")->with("error",$Action->errors);
-            
-        }
-        else {
-            if (!empty($request->get("idAction"))) {
-                $id = $Action->saveAction($request);
-                
+        if (Auth::User()->type == "user") {
+            $Action = new Action;
+            $Action->checkAddActionDate($request);
+            $Action->checkAddAction($request);
+            //$Action->checkAddMood($request);
+            if (count($Action->errors) != 0) {
+                return View("ajax.error")->with("error",$Action->errors);
+
             }
-            
+            else {
+                if (!empty($request->get("idAction"))) {
+                    $id = $Action->saveAction($request);
+
+                }
+
+            }
         }
-        
     }
     public function SleepEdit(Request $request) {
-        $awek = Sleep::selectAwek($request->get("id"));
-        return View("ajax.editSleep")->with("awek",$awek->how_wake_up)->with("id",$request->get("id"))->with("i",$request->get("i"));
+        if (Auth::User()->type == "user") {
+            $awek = Sleep::selectAwek($request->get("id"));
+            return View("ajax.editSleep")->with("awek",$awek->how_wake_up)->with("id",$request->get("id"))->with("i",$request->get("i"));
+        }
     }
     public function ShowDescription(Request $request) {
-        $description = (AppMood::showDescription($request->get("id")));
+        $description = (AppMood::showDescription($request->get("id"),Auth::User()->id));
         return View("ajax.description")->with("description",$description->what_work);
     }
     public function SleepEditAction(Request $request) {
-        $Mood = new Mood;
-        $Mood->IfInt($request->get("sleep")," ilośc wybudzen ");
-        if (count($Mood->errors) > 0 ) {
-            return View("ajax.error")->with("error",$Mood->errors);
-        }
-        else {
-            $Mood->updateSleep($request);
-            return View("ajax.succes")->with("succes","Pomyslnie edytowano");
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+            $Mood->IfInt($request->get("sleep")," ilośc wybudzen ");
+            if (count($Mood->errors) > 0 ) {
+                return View("ajax.error")->with("error",$Mood->errors);
+            }
+            else {
+                $Mood->updateSleep($request);
+                return View("ajax.succes")->with("succes","Pomyslnie edytowano");
+            }
         }
     }
     
     public function ActionShow(Request $request) {
         
         $Action = new Action;
-        $list = $Action->showListActionMood($request);
+        $list = $Action->showListActionMood($request,Auth::User()->id);
         return View("ajax.ActionShow")->with("list",$list);
         
          
@@ -147,11 +161,15 @@ class MoodController extends Controller  {
     
     
     public function delete(Request $request) {
-        $Mood = new Mood;
-        $Mood->deleteMood($request);
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+            $Mood->deleteMood($request);
+        }
     }
     public function SleepDelete(Request $request) {
-        $Mood = new Mood;
-        $Mood->deleteSleep($request);
+        if (Auth::User()->type == "user") {
+            $Mood = new Mood;
+            $Mood->deleteSleep($request);
+        }
     }
 }
